Application would slow on high resource consumption. Introducing interpolation to the result physics bodys' positions, making performance much smoother. Source for integration -> https://www.unagames.com/blog/daniele/2010/06/fixed-time-step-implementation-box2d

+++ include/gameworld.h
@@ -25,5 +25,5 @@ public:
 
 	void InitTestArea();
 
-	bool GameLoop(double &accumulator);
+	bool GameLoop(float & accumulator);
 };
